CHE-4462; refactory auth modules
