+++ plugins/plugin-microsoft/codenvy-plugin-microsoft-git-provider/src/main/java/org/eclipse/che/ide/ext/git/server/nativegit/MicrosoftOAuthCredentialsProvider.java
@@ -14,16 +14,14 @@
  */
 package org.eclipse.che.ide.ext.git.server.nativegit;
 
-import org.eclipse.che.api.auth.oauth.OAuthTokenProvider;
 import org.eclipse.che.api.auth.shared.dto.OAuthToken;
+import org.eclipse.che.api.git.CredentialsProvider;
+import org.eclipse.che.api.git.UserCredential;
 import org.eclipse.che.api.git.exception.GitException;
 import org.eclipse.che.api.git.shared.ProviderInfo;
 import org.eclipse.che.commons.annotation.Nullable;
 import org.eclipse.che.commons.env.EnvironmentContext;
-
-import org.eclipse.che.api.git.CredentialsProvider;
-import org.eclipse.che.api.git.UserCredential;
-
+import org.eclipse.che.security.oauth.shared.OAuthTokenProvider;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
